Added support for "score_threshold" and "max_num" for NMS operation when exporting to ONNX.
These parameters are supported in ONNX: [link](https://github.com/onnx/onnx/blob/master/docs/Operators.md#NonMaxSuppression).

In addition, these changes significantly speed up the export time:
|  Model name | Export time before (s) | Export time now (s) |
|:-----------:|:----------------------:|:----------------------:|
| ssd_300     |         1201.27        |          6.85          |
| yolo_v3     |          88.89         |          10.46         |
| fsaf        |          19.89         |          12.62         |
| retinanet   |          59.33         |          13.30         |
| faster_rcnn |          21.24         |          18.02         |

In addition, performance is improved in ONNX Runtime.
On the example of the RetinaNet model, the changes increase the FPS from 0.01 to 3.15.
Model profiling results in ONNX Runtime:
|                   | Befor (ms) | Now (ms) |
|-------------------|------------|----------|
| NonMaxSuppression | 59817.752  | 0.385    |